# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ABSuite API
  version: 1.0.0
  description: |
    Scoped, expiring, auditable credentials and execution infrastructure for AI agents.

    ## Authentication

    Three credential types, used for different purposes:

    | Header | Purpose |
    |---|---|
    | `Authorization: Bearer <capability-token>` | Normal service calls. Must carry the scope the endpoint requires. |
    | `X-ABSuite-Admin-Key` | Bootstrap credential. Grants full authority; used to mint the first token and manage tenants. |
    | `X-ABSuite-Tenant-Key` | Identifies the billed tenant. Optional — omit for an unmetered self-hosted deployment. |

    ## Scopes

    Scopes are `resource:action`, matched segment-wise. `read:*` grants
    `read:users` but never `read:users:delete`. A bare `*` grants everything.

    ## Quotas

    When a tenant key is supplied, plan limits are enforced. Exceeding a limit
    returns `402`; a suspended account returns `403`.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8081
    description: CapKit
  - url: http://localhost:8082
    description: Edge-Run
  - url: http://localhost:8083
    description: QuickBench
  - url: http://localhost:8084
    description: Connector-Starter
...
components:
  securitySchemes:
    capabilityToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A CapKit capability token carrying the required scope.
    adminKey:
      type: apiKey
      in: header
      name: X-ABSuite-Admin-Key
    tenantKey:
      type: apiKey
      in: header
      name: X-ABSuite-Tenant-Key
  schemas:
...
    TaskDefinition:
      oneOf:
        - type: object
          required: [type, url]
          properties:
            type: { type: string, enum: [http] }
            url: { type: string, format: uri }
            method: { type: string, default: GET }
            headers: { type: object, additionalProperties: { type: string } }
            body: {}
        - type: object
          required: [type, script]
          description: Only available when EDGERUN_SCRIPT_ROOT is configured.
          properties:
            type: { type: string, enum: [script] }
            script: { type: string }
            args:
              type: array
              items: { type: string }
paths:
...
  /schedule:
    post:
      tags: [Scheduling]
      summary: Create a cron schedule
      description: Requires scope `schedule:create`. Edge-Run, port 8082.
      security:
        - capabilityToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, cron, task]
              properties:
                id: { type: string }
                cron: { type: string, example: '*/15 * * * *' }
                task: { $ref: '#/components/schemas/TaskDefinition' }
                retry:
                  type: object
                  properties:
                    maxAttempts: { type: integer, default: 3 }
                    backoff: { type: string, enum: [fixed, exponential] }
                    baseDelay: { type: integer }
                timeout: { type: integer }
      responses:
        '201':
          description: Scheduled
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  nextRun: { type: string, format: date-time }
                  durable: { type: boolean }
    get:
      tags: [Scheduling]
      summary: List schedules
      description: Requires scope `schedule:read`.
      security:
        - capabilityToken: []
      responses:
        '200': { description: Registered schedules }
  /queue:
    post:
      tags: [Scheduling]
      summary: Enqueue a task
      description: Requires scope `queue:write`. Edge-Run, port 8082.
      security:
        - capabilityToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [task]
              properties:
                id: { type: string }
                priority: { type: string, enum: [high, normal, low], default: normal }
                delay: { type: integer, description: Milliseconds before it becomes runnable }
                task: { $ref: '#/components/schemas/TaskDefinition' }
      responses:
        '201': { description: Queued }
        '429': { description: Queue limit reached }
...
  /connectors/{id}/verify:
    post:
      tags: [Connectors]
      summary: Verify connector credentials
      description: |
        Requires scope `connector:read`. Read-only: verification never posts a
        message or creates a record as a side effect.
      security:
        - capabilityToken: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Credentials are valid }
        '503': { description: Not configured or rejected by the provider }
  /generate:
    post:
      tags: [Connectors]
      summary: Generate a connector from a description
      description: |
        Deterministic and rule-based — no API key, and the same description
        always produces identical output.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt: { type: string, example: Read GitHub issues and post them to Slack every 15 minutes }
      responses:
        '200':
          description: Generated connector
          content:
            application/json:
              schema:
                type: object
                properties:
                  manifest: { type: string, description: YAML manifest }
                  typescript: { type: string, description: Compilable connector source }
                  spec: { type: object }